fix: 코드 분할 및 컨펌 전 초안
